requirements.txt
-f /usr/share/pip-wheels
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.7
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap4==22.3
django-bower==5.2.0
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-formtools==2.3
django-jquery==3.1.0
django-npm==1.0.0
django-tempus-dominus==5.1.2.17
djangorestframework==3.12.4
django-q==1.3.9
djongo==1.3.6
dnspython==2.1.0
frozenlist==1.3.3
idna==3.2
mega.py==1.0.8
multidict==6.0.4
openai==0.27.0
pathlib==1.0.1
pexels-api==1.0.1
Pillow==9.3.0
pycryptodome==3.13.0
pydantic==1.10.4
pymongo==3.12.0
pytz==2021.1
requests==2.26.0
six==1.16.0
social-post-api==1.1.1
sorl-thumbnail==12.9.0
soupsieve==2.3.1
sqlparse==0.2.4
tenacity==5.1.5
tqdm==4.65.0
typing-extensions==4.4.0
urllib3==1.26.6
Wikipedia-API==0.5.4
yarl==1.8.2
